pyproject.toml
[project]
name = "django-service-layer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-rest-framework>=0.1.0",
]
